Use node object to determine azure resource
Instead of just passing along a node name, use the actual k8s node
object which has the provider ID associated with it. This allows us to
determine, per node, if it is a VM or part of a VMSS and handle
accordingly.

It also allows us to get the real resource name, especially in the case
of VMSS where we'd otherwise just have the node name.