Add support for vmss (#120)
* Add support for vmss

This adds support for assigning identities to a VMSS when the underlying
cluster is using VMSS.

Note that identities get assigned to the entire VMSS.

* Add vmss test client and tests

* Use node object to determine azure resource

Instead of just passing along a node name, use the actual k8s node
object which has the provider ID associated with it. This allows us to
determine, per node, if it is a VM or part of a VMSS and handle
accordingly.

It also allows us to get the real resource name, especially in the case
of VMSS where we'd otherwise just have the node name.