Fixes #137 - Remove ambiguity in naming requirements
I've found that if you choose a Managed Identity resource name that is different from the AzureIdentity name (aadpodidentity.yaml), the AzureIdentityBinding (aadpodidentitybinding.yaml) must reference the AzureIdentity name rather than the MSI resource name.  Use of `<any-idname>` in the aadpodidentity.yaml example confused me in this scenario, so suggesting this change.