Change TestEventRecorder buffer size to prevent blocking
Normally I would not just randomly bump up a channel like this, however
given this is for testing purposes and this implementation is
specifically for testing, I think it's fine.

Without this change we can end up with the channel send blocking while
waiting for the mic client's Sync from exiting.