Adds support for MIC to authenticate with azure using system assigned/user assigned MSI (#265)
* Adds support for MIC to authenticate with azure using system assigned or user assigned MSI.
Resolves the item in #261.
This PR adds the capability for MIC to look at azure.json or environment variables
to determine whether the system assigned or user assigned MSI has to be used for accessing
azure resources. The MIC requests for token based on MSI. Also contains changes in NMI to determine
if the request is originating from an MIC replicaset. If so, NMI directly generates the tokens
instead of looking up the azure assigned identity for the pod-binding match.

* Changes to accomodate merge from head of the tree

* Fix merge issues

* Address code reviews

* Refactor code to functions

* Simplify code

* README instructions

* README instructions - iteration 1

* README instructions - update

* README instructions - update

* README instructions - update

* Fix the identity remove from VMSS for system assigned identity only scenarios.

* Doc updates

* Update logging and fix the isMIC parameters

* Add adal TODO

* Address review comments.