feat: option to set Retry-After in NMI responses
Adds a new feature flag to enable setting Retry-After header in the
error response from NMI. The error is only when the identity is still
being assigned by NMI or no valid AzureAssignedIdentity is found yet.
This enables SDK's to retry based on the http status code 503 and the
retry after header.

Note: When enabling this feature, the default retries in NMI should be
explicitly disabled to rather rely on the SDK for retries.

--retry-attempts-for-created=1, --retry-attempts-for-assigned=1 and
--find-identity-retry-interval=1. This will force NMI to return
immediately with the Retry-After header 20s.

Signed-off-by: Anish Ramasekar <anish.ramasekar@gmail.com>